Added dev helpers to get preview/test data into the receiver
It looks like it will not be feasibile to directly call the HTTP trigger
for a Google-owned project due to policy which stops allUsers being able
to invoke. I have confirmed that it works fine on a non-Google owned
account so have just added in a helper to allow easy copy/pasting of a
command to submit responses in dev mode.

Also: add a CreatedAt column for the response table so we know when we
can easily monitor test responses as they come in.

Change-Id: I0408f1a750da057dcbfc3923b3f49064ee9bae10